+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Move loading of PostgreSQLProxy to its own Railtie
 - Move loading of Mysql2Proxy to its own Railtie
 - Move loading of TrilogyProxy to its own Railtie
+- Move loading of SQLite3Proxy to its own Railtie
 
 ## [0.8.0] - 2025-08-28